isolate migrate old queue to a different file
