+++ PostHog/PostHogFileBackedQueue.swift
@@ -27,7 +27,7 @@ class PostHogFileBackedQueue {
         }
 
         if oldQueue != nil {
-            migrateOldQueue(oldQueue!)
+            migrateOldQueue(queue: queue, oldQueue: oldQueue!)
         }
 
         do {
@@ -36,52 +36,6 @@ class PostHogFileBackedQueue {
         } catch {
             hedgeLog("Failed to load files for queue")
             // failed to read directory – bad permissions, perhaps?
-        }
-    }
-
-    private func migrateOldQueue(_ oldQueue: URL) {
-        if !FileManager.default.fileExists(atPath: oldQueue.path) {
-            return
-        }
-
-        var deleteFiles = false
-        defer {
-            if deleteFiles {
-                try? FileManager.default.removeItem(at: oldQueue)
-            }
-        }
-
-        do {
-            let data = try Data(contentsOf: oldQueue)
-            let array = try JSONSerialization.jsonObject(with: data) as? [Any]
-
-            if array == nil {
-                deleteFiles = true
-                return
-            }
-
-            for item in array! {
-                guard let event = item as? [String: Any] else {
-                    continue
-                }
-                let timestamp = event["timestamp"] as? String ?? toISO8601String(Date())
-
-                let timestampDate = toISO8601Date(timestamp) ?? Date()
-
-                let filename = "\(timestampDate.timeIntervalSince1970)"
-
-                let contents = try? JSONSerialization.data(withJSONObject: event)
-
-                if contents == nil {
-                    continue
-                }
-                try? contents!.write(to: queue.appendingPathComponent(filename))
-
-                deleteFiles = true
-            }
-        } catch {
-            deleteFiles = false
-            return
         }
     }
